Rename UI components folder to ui/
