+++ next/pages/auth/register.tsx
@@ -6,9 +6,9 @@ import { HeadingPage } from "@/components/heading--page";
 import { Meta } from "@/components/meta";
 import { getCommonPageProps } from "@/lib/get-common-page-props";
 
-import { Button } from "@/wunder-component-library/button";
-import { Input } from "@/wunder-component-library/input";
-import { StatusMessage } from "@/wunder-component-library/status-message";
+import { Button } from "@/ui/button";
+import { Input } from "@/ui/input";
+import { StatusMessage } from "@/ui/status-message";
 
 type Inputs = {
   name: string;
